# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User Profile API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/services": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Get list of all services",
                "description": "Retrieve a list of all services.",
                "operationId": "b2cc16bd5971e5680a4dd46c9c383f65",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Services"
                ],
                "summary": "Create a new service",
                "description": "Store a new service in the database.",
                "operationId": "806d2b381734b9494d468322434ba121",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "service_name": {
                                        "type": "string",
                                        "example": "Haircut"
                                    },
                                    "minutesAvailable": {
                                        "type": "string",
                                        "example": "30"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 20
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Service created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/services/{id}": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Get a specific service",
                "description": "Retrieve the details of a specific service.",
                "operationId": "6b8504a5f57dba118512a83b265e6040",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the service to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service details"
                    },
                    "404": {
                        "description": "Service not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Services"
                ],
                "summary": "Update an existing service",
                "description": "Update the details of an existing service.",
                "operationId": "a077666135d725a905538f3f05b4039c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the service to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "service_name": {
                                        "type": "string",
                                        "example": "Haircut"
                                    },
                                    "minutesAvailable": {
                                        "type": "string",
                                        "example": "30"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 20
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Service updated successfully"
                    },
                    "404": {
                        "description": "Service not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Services"
                ],
                "summary": "Delete a specific service",
                "description": "Permanently delete a service from the database.",
                "operationId": "66aaf62153671736305fd4cdd76702fc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the service to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service deleted successfully"
                    },
                    "404": {
                        "description": "Service not found"
                    }
                }
            }
        },
...
    },
...
}